pyproject.toml
[tool.poetry]
name = "pymisp"
version = "2.4.131"
description = "Python API for MISP."
authors = ["Raphaël Vinot <raphael.vinot@circl.lu>"]
license = "BSD-2-Clause"
repository = "https://github.com/MISP/PyMISP"
documentation = "http://pymisp.readthedocs.io"


readme = "README.md"

classifiers=[
    'License :: OSI Approved :: BSD License',
    'Development Status :: 5 - Production/Stable',
    'Environment :: Console',
    'Operating System :: POSIX :: Linux',
    'Intended Audience :: Science/Research',
    'Intended Audience :: Telecommunications Industry',
    'Intended Audience :: Information Technology',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Topic :: Security',
    'Topic :: Internet'
]

include = [
    "CHANGELOG.txt",
    "README.md",
    "pymisp/data/*.json",
    "pymisp/data/misp-objects/schema_objects.json",
    "pymisp/data/misp-objects/schema_relationships.json",
    "pymisp/data/misp-objects/objects/*/definition.json",
    "pymisp/data/misp-objects/relationships/definition.json",
    "pymisp/tools/pdf_fonts/Noto_TTF/*"
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/MISP/PyMISP/issues"
"Source" = "https://github.com/MISP/PyMISP"

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.22.0"
python-dateutil = "^2.8.1"
jsonschema = "^3.2.0"
deprecated = "^1.2.7"
python-magic = {version = "^0.4.15", optional = true}
pydeep = {version = "^0.4", optional = true}
lief = {version = "^0.10.1", optional = true}
beautifulsoup4 = {version = "^4.8.2", optional = true}
validators = {version = "^0.14.2", optional = true}
sphinx-autodoc-typehints = {version = "^1.10.3", optional = true}
recommonmark = {version = "^0.6.0", optional = true}
reportlab = {version = "^3.5.34", optional = true}
pyfaup = {version = "^1.2", optional = true}

[tool.poetry.extras]
fileobjects = ['python-magic', 'pydeep', 'lief']
openioc = ['beautifulsoup4']
virustotal = ['validators']
docs = ['sphinx-autodoc-typehints', 'recommonmark']
pdfexport = ['reportlab']
url = ['pyfaup']

[tool.poetry.dev-dependencies]
nose = "^1.3.7"
coveralls = "^1.11.1"
codecov = "^2.0.15"
requests-mock = "^1.7.0"
mypy = "^0.761"
flake8 = "^3.7.9"
ipython = "^7.12.0"
jupyterlab = "^1.2.6"

[build-system]
requires = ["poetry_core>=1.0", "setuptools"]
build-backend = "poetry.core.masonry.api"
